backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.3
Pillow>=10.0
gunicorn>=21.2
psycopg2-binary>=2.9
python-decouple>=3.8
mysqlclient>=2.1.1
whitenoise>=6.5
